@@ -1,6 +1,6 @@
 import json
+from collections.abc import Callable
 from datetime import timezone
-from typing import Optional
 
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
@@ -13,6 +13,7 @@ from django.db.models import (
 )
 from django.utils import timezone as django_timezone
 from django.utils.encoding import smart_str
+from django.utils.module_loading import import_string
 
 
 def track_field(field):
@@ -57,7 +58,7 @@ def get_fields_in_model(instance):
     return [f for f in instance._meta.get_fields() if track_field(f)]
 
 
-def get_field_value(obj, field, skip_json_dumps=False):
+def get_field_value(obj, field, use_json_for_changes=False):
     """
     Gets the value of a given model instance field.
 
@@ -68,11 +69,35 @@ def get_field_value(obj, field, skip_json_dumps=False):
     :return: The value of the field as a string.
     :rtype: str
     """
+
+    def get_default_value():
+        """
+        Attempts to get the default value for a field from the model's field definition.
+
+        :return: The default value of the field or None
+        """
+        try:
+            model_field = obj._meta.get_field(field.name)
+            default = model_field.default
+        except AttributeError:
+            default = NOT_PROVIDED
+
+        if default is NOT_PROVIDED:
+            default = None
+        elif callable(default):
+            default = default()
+
+        return smart_str(default) if not use_json_for_changes else default
+
     try:
         if isinstance(field, DateTimeField):
             # DateTimeFields are timezone-aware, so we need to convert the field
             # to its naive form before we can accurately compare them for changes.
-            value = field.to_python(getattr(obj, field.name, None))
+            value = getattr(obj, field.name)
+            try:
+                value = field.to_python(value)
+            except TypeError:
+                return value
             if (
                 value is not None
                 and settings.USE_TZ
@@ -80,27 +105,59 @@ def get_field_value(obj, field, skip_json_dumps=False):
             ):
                 value = django_timezone.make_naive(value, timezone=timezone.utc)
         elif isinstance(field, JSONField):
-            value = field.to_python(getattr(obj, field.name, None))
-            if not skip_json_dumps:
-                value = json.dumps(value, sort_keys=True, cls=field.encoder)
-        elif (field.one_to_one or field.many_to_one) and hasattr(
-            field, "rel_class"
-        ):
-            value = smart_str(
-                getattr(obj, field.get_attname(), None), strings_only=True
-            )
+            value = field.to_python(getattr(obj, field.name))
+            if not use_json_for_changes:
+                try:
+                    value = json.dumps(value, sort_keys=True, cls=field.encoder)
+                except TypeError:
+                    pass
+        elif (field.one_to_one or field.many_to_one) and hasattr(field, "rel_class"):
+            value = smart_str(getattr(obj, field.get_attname()), strings_only=True)
         else:
-            value = smart_str(getattr(obj, field.name, None))
-            if type(value).__name__ == "__proxy__":
-                value = str(value)
-    except ObjectDoesNotExist:
-        value = (
-            field.default
-            if getattr(field, "default", NOT_PROVIDED) is not NOT_PROVIDED
-            else None
-        )
+            value = getattr(obj, field.name)
+            if not use_json_for_changes:
+                value = smart_str(value)
+                if type(value).__name__ == "__proxy__":
+                    value = str(value)
+    except (ObjectDoesNotExist, AttributeError):
+        return get_default_value()
 
     return value
+
+
+def is_primitive(obj) -> bool:
+    """
+    Checks if the given object is a primitive Python type that can be safely serialized to JSON.
+
+    :param obj: The object to check
+    :return: True if the object is a primitive type, False otherwise
+    :rtype: bool
+    """
+    primitive_types = (type(None), bool, int, float, str, list, tuple, dict, set)
+    return isinstance(obj, primitive_types)
+
+
+def get_mask_function(mask_callable: str | None = None) -> Callable[[str], str]:
+    """
+    Get the masking function to use based on the following priority:
+    1. Model-specific mask_callable if provided
+    2. mask_callable from settings if configured
+    3. Default mask_str function
+
+    :param mask_callable: The dotted path to a callable that will be used for masking.
+    :type mask_callable: str
+    :return: A callable that takes a string and returns a masked version.
+    :rtype: Callable[[str], str]
+    """
+
+    if mask_callable:
+        return import_string(mask_callable)
+
+    default_mask_callable = settings.AUDITLOG_MASK_CALLABLE
+    if default_mask_callable:
+        return import_string(default_mask_callable)
+
+    return mask_str
 
 
 def mask_str(value: str) -> str:
@@ -117,7 +174,10 @@ def mask_str(value: str) -> str:
 
 
 def model_instance_diff(
-    old: Optional[Model], new: Optional[Model], fields_to_check=None
+    old: Model | None,
+    new: Model | None,
+    fields_to_check=None,
+    use_json_for_changes=False,
 ):
     """
     Calculates the differences between two model instances. One of the instances may be ``None``
@@ -130,6 +190,8 @@ def model_instance_diff(
     :type new: Model
     :param fields_to_check: An iterable of the field names to restrict the diff to, while ignoring the rest of
         the model's fields. This is used to pass the `update_fields` kwarg from the model's `save` method.
+    :param use_json_for_changes: whether or not to use a JSON for changes
+        (see settings.AUDITLOG_STORE_JSON_CHANGES)
     :type fields_to_check: Iterable
     :return: A dictionary with the names of the changed fields as keys and a two tuple of the old and new
             field values as value.
@@ -206,10 +268,10 @@ def model_instance_diff(
 
     for field in fields:
         old_value = get_field_value(
-            old, field, field.name in custom_fields_callbacks
+            old, field, use_json_for_changes or field.name in custom_fields_callbacks
         )
         new_value = get_field_value(
-            new, field, field.name in custom_fields_callbacks
+            new, field, use_json_for_changes or field.name in custom_fields_callbacks
         )
 
         if old_value != new_value:
@@ -225,12 +287,24 @@ def model_instance_diff(
                     custom_diff_result,
                 )
             elif model_fields and field.name in model_fields["mask_fields"]:
+                mask_func = get_mask_function(model_fields.get("mask_callable"))
                 diff[field.name] = (
-                    mask_str(smart_str(old_value)),
-                    mask_str(smart_str(new_value)),
+                    mask_func(smart_str(old_value)),
+                    mask_func(smart_str(new_value)),
                 )
             else:
-                diff[field.name] = (smart_str(old_value), smart_str(new_value))
+                if not use_json_for_changes:
+                    diff[field.name] = (smart_str(old_value), smart_str(new_value))
+                else:
+                    # TODO: should we handle the case where the value is a django Model specifically?
+                    #       for example, could create a list of ids for ManyToMany fields
+
+                    # this maintains the behavior of the original code
+                    if not is_primitive(old_value):
+                        old_value = smart_str(old_value)
+                    if not is_primitive(new_value):
+                        new_value = smart_str(new_value)
+                    diff[field.name] = (old_value, new_value)
 
     if len(diff) == 0:
         diff = None
