@@ -125,7 +125,13 @@ class Command(BaseCommand):
         def postgres():
             with connection.cursor() as cursor:
                 cursor.execute(
-                    'UPDATE auditlog_logentry SET changes="changes_text"::jsonb'
+                    """
+                    UPDATE auditlog_logentry
+                    SET changes="changes_text"::jsonb
+                    WHERE changes_text IS NOT NULL
+                        AND changes_text <> ''
+                        AND changes IS NULL
+                    """
                 )
                 return cursor.cursor.rowcount
 
