@@ -1,11 +1,9 @@
 # Generated by Django 4.0 on 2022-08-04 15:41
-from typing import List
-
 from django.conf import settings
 from django.db import migrations, models
 
 
-def two_step_migrations() -> List:
+def two_step_migrations() -> list:
     if settings.AUDITLOG_TWO_STEP_MIGRATION:
         return [
             migrations.RenameField(
