@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AuditlogTestConfig(AppConfig):
-    name = "auditlog_tests"
+    name = "test_app"
