@@ -8,6 +8,8 @@ DEBUG = True
 
 SECRET_KEY = "test"
 
+TEST_DB_BACKEND = os.getenv("TEST_DB_BACKEND", "sqlite3")
+
 INSTALLED_APPS = [
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -15,9 +17,11 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.admin",
     "django.contrib.staticfiles",
+    "django.contrib.postgres",
     "auditlog",
-    "auditlog_tests",
+    "test_app",
 ]
+
 
 MIDDLEWARE = [
     "django.middleware.common.CommonMiddleware",
@@ -27,18 +31,52 @@ MIDDLEWARE = [
     "auditlog.middleware.AuditlogMiddleware",
 ]
 
-DATABASES = {
-    "default": {
-        "ENGINE": "django.db.backends.postgresql",
-        "NAME": os.getenv(
-            "TEST_DB_NAME", "auditlog" + os.environ.get("TOX_PARALLEL_ENV", "")
-        ),
-        "USER": os.getenv("TEST_DB_USER", "postgres"),
-        "PASSWORD": os.getenv("TEST_DB_PASS", ""),
-        "HOST": os.getenv("TEST_DB_HOST", "127.0.0.1"),
-        "PORT": os.getenv("TEST_DB_PORT", "5432"),
+if TEST_DB_BACKEND == "postgresql":
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.postgresql",
+            "NAME": os.getenv(
+                "TEST_DB_NAME", "auditlog" + os.environ.get("TOX_PARALLEL_ENV", "")
+            ),
+            "USER": os.getenv("TEST_DB_USER", "postgres"),
+            "PASSWORD": os.getenv("TEST_DB_PASS", ""),
+            "HOST": os.getenv("TEST_DB_HOST", "127.0.0.1"),
+            "PORT": os.getenv("TEST_DB_PORT", "5432"),
+        }
     }
-}
+elif TEST_DB_BACKEND == "mysql":
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.mysql",
+            "NAME": os.getenv(
+                "TEST_DB_NAME", "auditlog" + os.environ.get("TOX_PARALLEL_ENV", "")
+            ),
+            "USER": os.getenv("TEST_DB_USER", "root"),
+            "PASSWORD": os.getenv("TEST_DB_PASS", ""),
+            "HOST": os.getenv("TEST_DB_HOST", "127.0.0.1"),
+            "PORT": os.getenv("TEST_DB_PORT", "3306"),
+            "OPTIONS": {
+                "charset": "utf8mb4",
+                "init_command": "SET sql_mode='STRICT_TRANS_TABLES'",
+            },
+        }
+    }
+elif TEST_DB_BACKEND == "sqlite3":
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.sqlite3",
+            "NAME": os.getenv(
+                "TEST_DB_NAME",
+                (
+                    ":memory:"
+                    if os.getenv("TOX_PARALLEL_ENV")
+                    else "test_auditlog.sqlite3"
+                ),
+            ),
+        }
+    }
+else:
+    raise ValueError(f"Unsupported database backend: {TEST_DB_BACKEND}")
 
 TEMPLATES = [
     {
@@ -57,7 +95,7 @@ TEMPLATES = [
 
 STATIC_URL = "/static/"
 
-ROOT_URLCONF = "auditlog_tests.urls"
+ROOT_URLCONF = "test_app.urls"
 
 USE_TZ = True
 
